fix(ci): python-coverage-shards — keep loop.py under the 1500-line maintainability limit
#657 pushed src/awf/runtime/pr_monitor_runner/loop.py to 1520 lines, tripping the hard
1500-line first-party file limit enforced by
test_core_decomposition_maintainability.test_first_party_code_files_stay_under_line_limit
(run inside the coverage shards). Move the module-level
_post_workflow_scope_notification_best_effort helper verbatim into a new sibling module
loop_helpers.py; behavior is unchanged and the three call sites in _execute() are intact.
